// CityBoard/cityboard-front/src/models/users.ts
/**
 * Generated by orval v6.11.1 🍺
 * Do not edit manually.
 * CityBoard API
 * OpenAPI spec version: v1
 */
import type { UserStatus } from './userStatus';
import type { Roles } from './roles';

export interface Users {
  id?: number;
  username?: string;
  password?: string;
  name?: string;
  surname?: string;
  middle_name?: string;
  status?: UserStatus;
  password_expired?: boolean;
  roles?: Roles[];
}