// CityBoard/cityboard-front/src/models/roles.ts
/**
 * Generated by orval v6.11.1 🍺
 * Do not edit manually.
 * CityBoard API
 * OpenAPI spec version: v1
 */

export type Roles = typeof Roles[keyof typeof Roles];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const Roles = {
  ROLE_USER: 'ROLE_USER',
  ROLE_MOD: 'ROLE_MOD',
  ROLE_ADMIN: 'ROLE_ADMIN',
} as const;