// CityBoard/cityboard-front/src/models/userStatus.ts
/**
 * Generated by orval v6.11.1 🍺
 * Do not edit manually.
 * CityBoard API
 * OpenAPI spec version: v1
 */

export type UserStatus = typeof UserStatus[keyof typeof UserStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const UserStatus = {
  ACTIVE: 'ACTIVE',
  BANNED: 'BANNED',
  DELETED: 'DELETED',
} as const;